Fixed some warnings
Former-commit-id: b8b389d9b9c1a0ece212b2ce5fec1ad575d26184

+++ src/server/game/DataStores/DBCEnums.h
@@ -78,7 +78,7 @@ enum AreaFlags
 #endif
 };
 
-enum Difficulty : uint32
+enum Difficulty : uint8
 {
     REGULAR_DIFFICULTY = 0,
